fasthtml-styling-experiments/requirements.txt
python-fasthtml>=0.12.33
